style: move currency helpers out of extensions dir
