+++ packages/sealed_currencies/lib/sealed_currencies.dart
@@ -4,6 +4,6 @@ library sealed_currencies;
 export "package:sealed_languages/sealed_languages.dart";
 
 export "src/data/fiat_currencies.data.dart";
-export "src/helpers/extensions/fiat_currency/fiat_currency_format.dart";
-export "src/helpers/extensions/fiat_currency/fiat_currency_getters.dart";
+export "src/helpers/fiat_currency/fiat_currency_format.dart";
+export "src/helpers/fiat_currency/fiat_currency_getters.dart";
 export "src/model/currency.dart";
