+++ packages/sealed_currencies/lib/src/model/currency.dart
@@ -1,5 +1,5 @@
 import "../data/fiat_currencies.data.dart";
-import "../helpers/extensions/fiat_currency/fiat_currency_getters.dart";
+import "../helpers/fiat_currency/fiat_currency_getters.dart";
 
 part "submodels/crypto_currency.dart";
 part "submodels/fiat_currency.dart";
