+++ packages/sealed_currencies/test/src/helpers/fiat_currency/fiat_currency_formatting_extension_test.dart
@@ -1,5 +1,5 @@
-import "package:sealed_currencies/src/helpers/extensions/fiat_currency/fiat_currency_format.dart";
-import "package:sealed_currencies/src/helpers/extensions/fiat_currency/fiat_currency_getters.dart";
+import "package:sealed_currencies/src/helpers/fiat_currency/fiat_currency_format.dart";
+import "package:sealed_currencies/src/helpers/fiat_currency/fiat_currency_getters.dart";
 import "package:sealed_currencies/src/model/currency.dart";
 import "package:test/test.dart";
